ragbot/requirements.txt
langchain
langchain-community
langchain-groq
faiss-cpu
pypdf
sentence-transformers
python-dotenvfastapi
uvicorn